clase5/app/requirements.txt
Flask==2.1.1
flask-swagger-ui
pymysql==1.0.2
Werkzeug==2.1.2
mysql-connector-python==8.0.32
flask-restx==0.5.1